+++ src/codeLens.ts
@@ -34,75 +34,64 @@ import { childProcesses } from "./runOnMarkdown";
 export const blockRegex = () => /^(`{3,})(.*?)\n(.*?)^\1$/gms;
 
 // Parses blocks
-export function parseBlock(
-  document: vscode.TextDocument,
-  match: RegExpExecArray
-) {
-  const parsedLang = match[2].trim().toLowerCase();
-  const language = parsedLang === "" ? "bash" : parsedLang; // Treat untitled blocks as bash files
+export function parseBlock(doc: vscode.TextDocument, match: RegExpExecArray) {
+  const parsedLang = (match[2].trim().toLowerCase().match(/^\w+/) ?? [""])[0];
+  const language = !parsedLang ? "bash" : parsedLang; // Treat untitled blocks as bash files
   const code = match[3];
-  const start = document.positionAt(match.index);
-  const end = document.positionAt(match.index + match[0].length);
+  const start = doc.positionAt(match.index);
+  const end = doc.positionAt(match.index + match[0].length);
   const range = new vscode.Range(start, end);
   return { language, code, range };
 }
 
 // Generate the code lens with the required parameters and push it to the list
-function pushCodeLens(
+function add(
   lenses: vscode.CodeLens[],
   range: vscode.Range,
   title: string,
   command: string,
-  args: unknown[]
+  args: unknown[],
 ) {
   lenses.push(new vscode.CodeLens(range, { title, command, arguments: args }));
 }
 
 // Implementation for ButtonCodeLensProvider
-function provideCodeLenses(document: vscode.TextDocument): vscode.CodeLens[] {
+function provideCodeLenses(document: vscode.TextDocument) {
   const lenses: vscode.CodeLens[] = [];
 
   // Generate buttons to stop `Run on Markdown` processes
   for (const { pid, line } of childProcesses) {
     const range = new vscode.Range(line, 0, line, 0);
-    const argsStop = [pid, "SIGINT"];
-    const argsKill = [pid, "SIGKILL"];
-    pushCodeLens(lenses, range, `Stop`, `markdown.killProcess`, argsStop);
-    pushCodeLens(lenses, range, `Kill`, `markdown.killProcess`, argsKill);
+    add(lenses, range, "Stop", "markdown.killProcess", [pid, "SIGINT"]);
+    add(lenses, range, "Kill", "markdown.killProcess", [pid, "SIGKILL"]);
   }
 
   // Loop through all parsed code blocks and generate buttons
   for (const match of document.getText().matchAll(blockRegex())) {
     const { language, code, range } = parseBlock(document, match);
-    const name = getLanguageConfig(language, `name`);
+    const name = getLanguageConfig(language, "name");
 
     // For all supported languages, provide options to run the code block
     if (name !== undefined) {
       // Provide `Run {Language} Block` option
-      const titleRun = `Run ${name} Block`;
-      const commandRun = `markdown.runFile`;
       const argsRun = [language, code];
-      pushCodeLens(lenses, range, titleRun, commandRun, argsRun);
+      add(lenses, range, `Run ${name} Block`, "markdown.runFile", argsRun);
 
       // Only for bash code blocks, provide `Run in Terminal (line by line)` option
-      const titleBash = `Run in Terminal`;
-      const commandBash = `markdown.runInTerminal`;
-      if (language === `bash` || language === `powershell`)
-        pushCodeLens(lenses, range, titleBash, commandBash, [code]);
+      if (language === "bash" || language === "powershell")
+        add(lenses, range, "Run in Terminal", "markdown.runInTerminal", [code]);
 
       // Provide `Run on Markdown` option
-      const titleMark = `Run on Markdown`;
-      const commandMark = `markdown.runOnMarkdown`;
       const argsMark = [language, code, range];
-      pushCodeLens(lenses, range, titleMark, commandMark, argsMark);
+      add(lenses, range, "Run on Markdown", "markdown.runOnMarkdown", argsMark);
     }
 
     // Always provide buttons to copy, clear or delete code blocks
-    pushCodeLens(lenses, range, `Copy`, `markdown.copy`, [code]);
-    const clr = new vscode.Range(range.start.line + 1, 0, range.end.line, 0);
-    pushCodeLens(lenses, range, `Clear`, `markdown.delete`, [clr]);
+    const clear = new vscode.Range(range.start.line + 1, 0, range.end.line, 0);
     const del = new vscode.Range(range.start.line, 0, range.end.line + 1, 0);
-    pushCodeLens(lenses, range, `Delete`, `markdown.delete`, [del]);
+    add(lenses, range, "Copy", "markdown.copy", [code]);
+    add(lenses, range, "Clear", "markdown.delete", [clear]);
+    add(lenses, range, "Delete", "markdown.delete", [del]);
   }
 
   return lenses;
@@ -110,5 +99,10 @@ function provideCodeLenses(document: vscode.TextDocument): vscode.CodeLens[] {
 
 // CodeLens buttons provider for parsed code blocks
 export class ButtonCodeLensProvider implements vscode.CodeLensProvider {
+  private _emitter = new vscode.EventEmitter<void>();
+  onDidChangeCodeLenses = this._emitter.event;
   provideCodeLenses = provideCodeLenses;
+
+  // Expose method to refresh CodeLenses on-demand
+  public refresh = () => this._emitter.fire();
 }